+++ src/index.ts
@@ -1,10 +1,10 @@
 export { default as persistReducer } from './persistReducer';
 export { default as persistStore } from './persistStore';
 export { default as createMigrate } from './createMigrate';
-export { default as createTransform } from './createTransform';
 export { default as getStoredState } from './getStoredState';
 export { default as createPersistoid } from './createPersistoid';
 export { default as purgeStoredState } from './purgeStoredState';
+export { createTransform } from './transforms';
 
 export * from './constants';
 
@@ -15,10 +15,6 @@ export type {
   Persistor,
   PersistorState,
   Persistoid,
-  Transform,
-  TransformConfig,
-  TransformInbound,
-  TransformOutbound,
   Storage,
   WebStorage,
   MigrationManifest,